Полный рефакторинг кода с переносом общей логики, которая будет использоваться в экранах ЕГЭ в модуль core
