+++ app/src/main/java/com/example/mytraining/precentation/navigation/NavigationBar.kt
@@ -23,15 +23,15 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.example.core.ui.theme.BottomNavigationColor
+import com.example.core.precentation.theme.BottomNavigationColor
 import com.example.feature_favorites.presentation.ui.FavoriteScreen
 import com.example.feature_oge.presentation.ui.AllQuestionsScreen
 import com.example.feature_oge.presentation.ui.SelectionTrainingSectionScreen
-import com.example.feature_oge.presentation.ui.TaskSelectionTicketScreen
+import com.example.core.precentation.ui.TaskSelectionTicketScreen
 import com.example.feature_oge.presentation.ui.TheoryScreen
 import com.example.feature_oge.presentation.ui.TicketDetailsScreen
-import com.example.feature_oge.presentation.ui.TicketsScreen
-import com.example.feature_oge.presentation.ui.YearSelectionScreen
+import com.example.core.precentation.ui.TicketsScreen
+import com.example.core.precentation.ui.YearSelectionScreen
 import com.example.feature_profile.presentation.ui.ProfileScreen
 import com.example.feature_profile.presentation.ui.TarifScreen
 import com.example.mytraining.precentation.ui.HomeScreen
