+++ app/src/main/java/com/example/mytraining/precentation/ui/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import com.example.mytraining.precentation.navigation.NavigationBarExample
-import com.example.core.ui.theme.MyTrainingTheme
+import com.example.core.precentation.theme.MyTrainingTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
