+++ core/build.gradle.kts
@@ -71,4 +71,8 @@ dependencies {
 
     //Datastore
     implementation(libs.androidx.datastore.preferences)
+
+    //Coil
+    implementation(libs.coil.compose)
+    implementation(libs.play.services.location)
 }
