+++ core/src/main/java/com/example/core/precentation/components/AnswerOptionItem.kt
@@ -1,0 +1,44 @@
+package com.example.core.precentation.components
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.dimensionResource
+import com.example.core.R
+
+/**
+ * Отвечает за размещение Checkbox и варината ответа.
+ */
+
+@Composable
+fun AnswerOptionItem(
+    option: String,
+    checked: Boolean,
+    onCheckedChange: (Boolean) -> Unit,
+    enabled: Boolean,
+) {
+    Row(
+        modifier = Modifier.padding(vertical = dimensionResource(R.dimen.padding_8dp)),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Checkbox(
+            checked = checked,
+            onCheckedChange = onCheckedChange,
+            enabled = enabled
+        )
+        Spacer(modifier = Modifier.width(dimensionResource(R.dimen.padding_8dp)))
+        Text(
+            option,
+            style = MaterialTheme.typography.labelSmall,
+            color = Color.White
+        )
+    }
+}
