+++ core/src/main/java/com/example/core/precentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.core.ui.theme
+package com.example.core.precentation.theme
 
 import androidx.compose.ui.graphics.Color
