+++ core/src/main/java/com/example/core/precentation/theme/Shapes.kt
@@ -1,4 +1,4 @@
-package com.example.core.ui.theme
+package com.example.core.precentation.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
