+++ core/src/main/java/com/example/core/precentation/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.core.ui.theme
+package com.example.core.precentation.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
