+++ core/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="select_some_answer">Выберите несколько вариантов ответа:</string>
+</resources>
