+++ feature-oge/src/main/java/com/example/feature_oge/presentation/ui/SelectionTrainingSectionScreen.kt
@@ -19,10 +19,10 @@ import androidx.compose.ui.res.dimensionResource
 import com.example.core.R
 import com.example.core.domain.models.TrainingSection
 import com.example.core.mock.mockSubjects
-import com.example.core.ui.theme.BackgroundGradientBlue
-import com.example.core.ui.theme.BackgroundGradientGreen
-import com.example.feature_oge.presentation.ui.components.RowContent
-import com.example.feature_oge.presentation.ui.components.TopIconButtonAndText
+import com.example.core.precentation.theme.BackgroundGradientBlue
+import com.example.core.precentation.theme.BackgroundGradientGreen
+import com.example.core.precentation.components.RowContent
+import com.example.core.precentation.components.TopIconButtonAndText
 
 /**
  *
