+++ feature-profile/src/main/java/com/example/feature_profile/presentation/ui/ProfileScreen.kt
@@ -33,10 +33,10 @@ import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.LottieConstants
 import com.airbnb.lottie.compose.rememberLottieComposition
-import com.example.core.ui.theme.BackgroundGradientBlue
-import com.example.core.ui.theme.BackgroundGradientGreen
-import com.example.core.ui.theme.CardBackgroundGradientBrown
-import com.example.core.ui.theme.CardBackgroundGradientGreen
+import com.example.core.precentation.theme.BackgroundGradientBlue
+import com.example.core.precentation.theme.BackgroundGradientGreen
+import com.example.core.precentation.theme.CardBackgroundGradientBrown
+import com.example.core.precentation.theme.CardBackgroundGradientGreen
 import com.example.feature_profile.R
 
 /**
